ml/requirements.txt
numpy 
torch 
gymnasium
requests
python-dotenv